chore:修正外键的字符类型

+++ database/migrations/2019_11_06_153508_create_keywords_table.php
@@ -15,7 +15,7 @@ class CreateKeywordsTable extends Migration
     {
         Schema::create('keywords', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('rule_id')->default(0);
+            $table->unsignedBigInteger('rule_id')->default(0);
             $table->string('keyword', 64);
             $table->unsignedTinyInteger('match_type')->default(1)->comment('匹配规则: 1全匹配 2半匹配');
             $table->timestamps();
